+++ src/dwarf_parser.h
@@ -2,7 +2,9 @@
 
 #include "struct.h"
 #include "types.h"
+#include "util/result.h"
 
+#include <filesystem>
 #include <memory>
 #include <string>
 #include <vector>
@@ -15,6 +17,23 @@
 namespace bpftrace {
 
 class BPFtrace;
+
+class DwarfParseError : public ErrorInfo<DwarfParseError> {
+public:
+  static char ID;
+  void log(llvm::raw_ostream &OS) const override;
+
+  DwarfParseError(std::string &&msg) : msg_(std::move(msg)) {};
+  DwarfParseError() = default;
+
+  const std::string &msg() const
+  {
+    return msg_;
+  }
+
+private:
+  std::string msg_;
+};
 
 class Dwarf {
 public:
@@ -30,7 +49,16 @@ public:
   SizedType get_stype(const std::string &type_name) const;
   void resolve_fields(const SizedType &type) const;
 
+  Result<uint64_t> line_to_addr(const std::string &source_file,
+                                size_t line_num,
+                                size_t col_num = 0) const;
+
 private:
+  struct CuInfo {
+    std::filesystem::path source;
+    Dwarf_Die *die;
+  };
+
   Dwarf(BPFtrace *bpftrace, const std::string &file_path);
 
   std::vector<Dwarf_Die> function_param_dies(const std::string &function) const;
@@ -46,12 +74,18 @@ private:
 
   SizedType get_stype(Dwarf_Die &type_die, bool resolve_structs = true) const;
 
+  Result<CuInfo> get_cu_info(const std::string &source_file) const;
+
   static std::optional<Dwarf_Die> get_child_with_tagname(
       Dwarf_Die *die,
       int tag,
       const std::string &name);
   static std::vector<Dwarf_Die> get_all_children_with_tag(Dwarf_Die *die,
                                                           int tag);
+
+  static std::optional<std::filesystem::path> resolve_cu_path(
+      std::string_view cu_name,
+      std::string_view cu_comp_dir);
 
   Dwfl *dwfl = nullptr;
   Dwfl_Callbacks callbacks;
@@ -100,6 +134,14 @@ public:
 
   void resolve_fields(const SizedType &type __attribute__((unused))) const
   {
+  }
+
+  Result<uint64_t> line_to_addr(const std::string &source_file
+                                __attribute__((unused)),
+                                size_t line_num __attribute__((unused)),
+                                size_t col_num __attribute__((unused))) const
+  {
+    return make_error<DwarfParseError>();
   }
 };